+++ src/ctap/mod.rs
@@ -41,9 +41,11 @@ use self::command::{
     AuthenticatorVendorConfigureParameters, AuthenticatorVendorUpgradeParameters, Command,
 };
 use self::config_command::process_config;
-use self::credential_id::MAX_CREDENTIAL_ID_SIZE;
+use self::credential_id::{
+    decrypt_credential_id, encrypt_to_credential_id, MAX_CREDENTIAL_ID_SIZE,
+};
 use self::credential_management::process_credential_management;
-use self::crypto_wrapper::{decrypt_credential_source, encrypt_key_handle, PrivateKey};
+use self::crypto_wrapper::PrivateKey;
 use self::data_formats::{
     AuthenticatorTransport, CoseKey, CoseSignature, CredentialProtectionPolicy,
     EnterpriseAttestationMode, GetAssertionExtensions, PackedAttestationStatement,
@@ -807,8 +809,7 @@ impl CtapState {
         if let Some(exclude_list) = exclude_list {
             for cred_desc in exclude_list {
                 if storage::find_credential(env, &rp_id, &cred_desc.key_id, !has_uv)?.is_some()
-                    || decrypt_credential_source(env, cred_desc.key_id, &rp_id_hash, !has_uv)?
-                        .is_some()
+                    || decrypt_credential_id(env, cred_desc.key_id, &rp_id_hash, !has_uv)?.is_some()
                 {
                     // Perform this check, so bad actors can't brute force exclude_list
                     // without user interaction.
@@ -882,7 +883,7 @@ impl CtapState {
             storage::store_credential(env, credential_source)?;
             random_id
         } else {
-            encrypt_key_handle(env, &private_key, &rp_id_hash, cred_protect_policy)?
+            encrypt_to_credential_id(env, &private_key, &rp_id_hash, cred_protect_policy)?
         };
 
         let mut auth_data = self.generate_auth_data(env, &rp_id_hash, flags)?;
@@ -1072,7 +1073,7 @@ impl CtapState {
                 return Ok(credential);
             }
             let credential =
-                decrypt_credential_source(env, allowed_credential.key_id, rp_id_hash, !has_uv)?;
+                decrypt_credential_id(env, allowed_credential.key_id, rp_id_hash, !has_uv)?;
             if credential.is_some() {
                 return Ok(credential);
             }
